Add modular Rust GUI backends and platform bridge
