+++ src/gui_gtk.c
@@ -1,13 +1,11 @@
 #include "vim.h"
 
-#ifdef FEAT_GUI_GTK
-
-extern void rs_gui_run(void);
+#if defined(FEAT_GUI_GTK) && defined(FEAT_GUI_RUST)
+# include "gui_rust.h"
 
 int gui_mch_init(void)
 {
     rs_gui_run();
     return OK;
 }
-
 #endif
